Move SentryConnector to Shared (#326)
* :truck: Move SentryConnector to shared

* :bulb: Add module doc to SentryConnector

* :bulb: Document use of the reportError function

* 2.18.3-3
