+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "learning-object-service",
-  "version": "2.18.3-2",
+  "version": "2.18.3-3",
   "description": "Learning Object Microservice.",
   "main": "app.ts",
   "scripts": {
