+++ src/Changelogs/ChangelogDatastore.ts
@@ -1,6 +1,6 @@
 import { Db } from 'mongodb';
 import { COLLECTIONS } from '../drivers/MongoDriver';
-import { reportError } from '../drivers/SentryConnector';
+import { reportError } from '../shared/SentryConnector';
 import { ResourceError, ResourceErrorReason, ServiceError, ServiceErrorReason } from '../shared/errors';
 import { ChangeLogDocument } from '../shared/types/Changelog';
