+++ src/shared/SentryConnector.ts
@@ -1,3 +1,10 @@
+/**
+ * This module exports an error reporting function for use in error cases we cannot recover from.
+ * Depending on the environment, reported errors are directed to different locations. In development,
+ * errors are reported on the stderr stream, while in production they are sent to Sentry.
+ *
+ * @module shared/SentryConnector
+ */
 import * as Sentry from '@sentry/node';
 import * as express from 'express';
 
@@ -20,4 +27,11 @@ switch (environment) {
 
 export const sentryRequestHandler = Sentry.Handlers.requestHandler() as express.RequestHandler;
 export const sentryErrorHandler = Sentry.Handlers.errorHandler() as express.ErrorRequestHandler;
+
+/**
+ * Sends the provided error to the reporting service registered for the current environment.
+ *
+ * @example
+ * reportError(new Error('Oh no! We broke the API!'));
+ */
 export const reportError = _reportError;
